link, tag를 embeddable로 선언, entity 제거
