+++ src/main/kotlin/com/waffletoy/team1server/post/dto/LinkDto.kt
@@ -1,0 +1,10 @@
+package com.waffletoy.team1server.post.dto
+
+data class LinkDto(
+    val description: String,
+    val link: String,
+) {
+    companion object {
+        fun fromLink(link: Link): LinkDto = LinkDto(link.description, link.link)
+    }
+}
