+++ src/main/kotlin/com/waffletoy/team1server/post/dto/Tag.kt
@@ -1,0 +1,8 @@
+package com.waffletoy.team1server.post.dto
+
+import jakarta.persistence.Embeddable
+
+@Embeddable
+data class Tag(
+    val tag: String,
+)
