--- src/main/kotlin/com/waffletoy/team1server/post/persistence/TagEntity.kt
@@ -1,14 +1,0 @@
-package com.waffletoy.team1server.post.persistence
-
-import jakarta.persistence.*
-import java.util.*
-
-@Entity
-@Table(name = "tags")
-class TagEntity(
-    @Id
-    @Column(name = "ID", nullable = false)
-    open val id: String = UUID.randomUUID().toString(),
-    @Column(name = "TAG", nullable = false)
-    val tag: String,
-)
